requirements.txt
autogen-agentchat>=0.2.0
autogen-ext[openai,azure]>=0.2.0
python-dotenv>=1.0.0
qdrant-client>=1.6.0,<2.0.0
sentence-transformers==2.2.2
transformers==4.25.1
torch>=1.11.0,<2.0.0
tokenizers==0.13.2
huggingface-hub==0.12.1
scikit-learn>=1.0.2
numpy>=1.21.2,<2.0.0
tqdm>=4.65.0
h2==4.1.0
httpx[http2]>=0.20.0,<0.24.0
